chore: Fix test by using fixtures
