+++ tests/functional/idempotency/test_idempotency.py
@@ -5,9 +5,7 @@ import sys
 import jmespath
 import pytest
 from botocore import stub
-from jmespath import functions
 
-from aws_lambda_powertools.utilities.idempotency import DynamoDBPersistenceLayer
 from aws_lambda_powertools.utilities.idempotency.exceptions import (
     IdempotencyAlreadyInProgressError,
     IdempotencyInconsistentStateError,
@@ -664,16 +662,11 @@ def test_jmespath_with_powertools_json(persistence_store):
     assert result == persistence_store._generate_hash(expected_value)
 
 
-def test_custom_jmespath_function_overrides_builtin_functions():
-    class CustomFunctions(functions.Functions):
-        @functions.signature({"types": ["string"]})
-        def _func_echo_decoder(self, value):
-            return value
-
-    persistence_store = DynamoDBPersistenceLayer(
-        table_name="foo",
-        event_key_jmespath="powertools_json(data).payload",
-        jmespath_options={"custom_functions": CustomFunctions()},
-    )
+@pytest.mark.parametrize("persistence_store_with_jmespath_options", ["powertools_json(data).payload"], indirect=True)
+def test_custom_jmespath_function_overrides_builtin_functions(persistence_store_with_jmespath_options):
+    # GIVEN an persistence store with a custom jmespath_options
+    # AND use a builtin powertools custom function
     with pytest.raises(jmespath.exceptions.UnknownFunctionError, match="Unknown function: powertools_json()"):
-        persistence_store._get_hashed_idempotency_key({})
+        # WHEN calling _get_hashed_idempotency_key
+        # THEN raise unknown function
+        persistence_store_with_jmespath_options._get_hashed_idempotency_key({})
